refactored code to companieshouse standards
